requirements.txt
streamlit==1.41.1
snowflake-connector-python==3.12.4
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
requests==2.32.3
